# docs/reference/openapi.yaml
openapi: 3.0.3
info:
  title: QRES RaaS API
  description: |
    QRES exposes two HTTP API surfaces:
    - **P2P Status API** (default port 8080, set via `--port`): Real-time swarm status from the running node.
    - **Management API** (default port 3030, set via `[api] port` in config.toml): Daemon lifecycle, stats, and configuration.
  version: "21.0.0"
  license:
    name: MIT OR Apache-2.0
servers:
  - url: http://localhost:8080
    description: P2P Status API (--port flag)
  - url: http://localhost:3030
    description: Management API ([api] port in config.toml)
...
paths:
...
  /api/swarm/start:
    post:
      tags: [Management]
      summary: Start the P2P daemon
      description: Launches the swarm node with optional WAN mode and gossip interval.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRequest"
      responses:
        "200":
          description: Daemon started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "500":
          description: Failed to start
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/swarm/stop:
    post:
      tags: [Management]
      summary: Stop the P2P daemon
      responses:
        "200":
          description: Daemon stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "500":
          description: Failed to stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    StatusResponse:
      type: object
      properties:
        running:
          type: boolean
        pid:
          type: integer
          nullable: true
        metrics:
          type: object
          nullable: true
    StartRequest:
      type: object
      properties:
        wan:
          type: boolean
          description: Enable WAN mode (Kademlia DHT)
          default: false
        gossip_interval:
          type: integer
          format: int64
          description: Gossip interval in milliseconds
          default: 600
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...